==> 00_data_preprocessing.ipynb <==
# Data Preprocessing

This notebook builds a master dataframe of financial journals, clean the data, and export data to csv files.

from nltk.stem.snowball import SnowballStemmer
from sklearn.preprocessing import StandardScaler
import regex as re
import numpy as np
import pandas as pd
import warnings
warnings.filterwarnings('ignore')

## Load Dataset

# get a list of csv file in folder Dataset
dataset_p = 'Dataset/Journal_of_Financial_Economics_{}.csv'

# read each csv file into a Pandas dataframe
# and store the dataframes into a list
list_df = [pd.read_csv(dataset_p.format(y)) for y in range(2020,1973,-1)]

# display number of dataframes created
print("Number of Data Frames:", len(list_df))

# look at the first dataframe
list_df[0].head()

# merge all dataframes into a master dataframe
df_master = pd.concat(list_df, axis=0, ignore_index=True)

# drop rows that have missing values
df_master.dropna(axis=0, how='any',inplace=True)

# get the name of first author
df_master['author_first'] = df_master.authors.apply(lambda x: x[1:-1].split(',')[0][1:-1])

# duplicate Year
df_master['year_master'] = df_master.year.copy()

# compute the standard deviation for Year
df_master['year_std'] = df_master.year / df_master.year.std()

# apply standard scaler on Year
df_master['year'] = StandardScaler().fit_transform(df_master['year'].values.reshape(-1, 1))

# display the shape of df_masters
print("Data's shape: ", df_master.shape)

# look at df_master
df_master.head()

## Data Preprocessing & Cleaning

from nltk.corpus import stopwords
import re

stops = set(stopwords.words("english"))    # get unique English stopwords
regex = re.compile(r'[^\w\s]+')            # regular expression pattern
stemmer = SnowballStemmer("english")       # create a Snowball Stemmer object for stemming

# duplicate a subset of the master dataframe
df = df_master[['year_master', 'year', 'year_std', 'month', 'keywords', 'abstract']]
df.rename(columns={'year_master': 'Year', 'year': 'Year_Scaled', 
                   'year_std': 'Year_STD', 'month': 'Month', 'keywords': 'Keywords',
                   'abstract': 'Abstract'}, inplace=True)

# clean the abstract
df['Abstract_Cleaned'] = df.Abstract.map(lambda x: ' '.join([stemmer.stem(y) for y in x.split(' ')]))
df['Abstract_Cleaned'] = df.Abstract_Cleaned.map(lambda x: ' '.join([y for y in x.split(' ') if not y in stops]))
df['Abstract_Cleaned'] = df.Abstract_Cleaned.map(lambda x: ' '.join([regex.sub('', x) for x in x.split(' ')]))

# get the length of the abstract
df['Abstract Length'] = df.Abstract.str.len()

def stem_keywords(keywords):
    # create an empty list of keywords
    kw_list = []
    
    # iterate through each keyword in the list and apply stemming
    for kw in keywords[1:-1].split(', '):
        # stem each word in the  keyword
        stemmed_kw = ' '.join(stemmer.stem(word) for word in kw.split(' '))
        kw_list.append(stemmed_kw)
    
    return kw_list


# apply stemming to Keywords
df['Keywords_Cleaned'] = df['Keywords'].apply(stem_keywords)

# count number of keywords
df['Number of Keywords'] = df['Keywords_Cleaned'].map(lambda keywords: \
                                                      0 if str(keywords) == "['']" else len(keywords))

### Month

# display unique value of month
df['Month'].unique()

import calendar

def process_month(month):
    '''
    - Convert month name to numeric value
    - Set any invalid month value to -1
    '''
    
    # extract month from a string
    month_extracted = month.split('–')[-1]
    
    # assume month's value is invalid
    month_numeric = -1
    
    if month_extracted.isnumeric() and int(month_extracted) in np.arange(1, 13):
        # month_extracted has value between 1 and 12
        month_numeric = int(month_extracted)
    elif not month_extracted.isnumeric():
        # month_extracted have text value "January" to "December"
        month_numeric = list(calendar.month_name).index(month_extracted)
        
    return month_numeric

# clean month value
df['Month_Cleaned'] = df['Month'].map(process_month)

# take a look at data
df.head()

## Export Data

df_master.to_csv('data/data_master.csv', index=False)               # save master data
df.to_csv('data/data_cleaned.csv', index=False)                     # save cleaned data
