==> 02_feature_engineering.ipynb <==
# Feature Engineering

This notebook prepares data for model training.

import numpy as np
import pandas as pd
from sklearn.pipeline import Pipeline
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
from sklearn.preprocessing import Normalizer
from sklearn.decomposition import LatentDirichletAllocation
from sklearn.decomposition import TruncatedSVD
from mlxtend.preprocessing import DenseTransformer
from sklearn.cluster import KMeans
from tqdm import tqdm

## Load Data

# load pre-processed data
df = pd.read_csv('data/data_cleaned.csv')
df.head()

## Hyperparameter Tuning

from tqdm import tqdm
from sklearn.metrics import silhouette_score


lda_n_compo = [20, 30, 40, 50, 60]     # a list of LDA's number of components
lda_topic_word = [None, 15, 20, 25]    # a list of number of terms
svd = [8, 10, 12, 15]                  # a list of SVD's number of components
params_result = []

# iterate through every item in the lda_n_compo list
# run KMeans model for each selected set of parameters
for ldn in tqdm(lda_n_compo):
    for ldt in lda_topic_word: 
        for sn in svd:
            try:
                if sn >= ldn or ldt >= ldn:
                    continue
            except TypeError:
                pass

            # build a data pipeline
            p = [('vect', CountVectorizer()),
                 ('lda', LatentDirichletAllocation(n_components=ldn,random_state=42,topic_word_prior=ldt)),
                 ('tfidf', TfidfTransformer())]
            
            # include dimensionality reduction in the data pipeline
            if sn > 0:
                p.append(('svd', TruncatedSVD(n_components=sn)))
            p.append(('to_dense', DenseTransformer()))
            c_pipe = Pipeline(p)
            
            # create a vector representation of abstracts
            vec = c_pipe.fit_transform(df.Abstract_Cleaned)
            # normalize vector
            vec = Normalizer().fit_transform(vec)
            vec = pd.DataFrame(vec, index=None)
            vec = vec / vec.std()
            vec = vec.to_numpy()
            
            # build a KMeans model and compute Silhouette score
            model = KMeans(n_clusters=15,random_state=42)
            predicted = model.fit_predict(vec)
            score = silhouette_score(vec, predicted)
            params_result.append((ldn, ldt, sn, score))

# create a dataframe of the parameter tuning results
# the best model should have a highest Silhouette score
df_params = pd.DataFrame(params_result, columns=['LDA','LDA_topic_word','SVD','Score'])
df_params.sort_values(['Score'], ascending=False).head(10)

## Build Vector Representation of Abstracts

# build a data pipeline
pipeline = Pipeline([('vect', CountVectorizer()),                                 # bag-of-words
                     ('lda', LatentDirichletAllocation(n_components=50,           # topic modeling
                                                       random_state=42,
                                                       topic_word_prior=None)),
                     ('tfidf', TfidfTransformer()),
                     ('svd', TruncatedSVD(n_components=10, random_state=42)),     # components reduction
                     ('to_dense', DenseTransformer())])                           # data transform

# build data vector representation of abstracts
x_vector = pipeline.fit_transform(df.Abstract_Cleaned)

# normalize x_vector
x_vector = Normalizer().fit_transform(x_vector)
df_x_vector = pd.DataFrame(x_vector, index=None)
df_x_vector = df_x_vector / df_x_vector.std()
x_vector = df_x_vector.to_numpy()

# display x_vector
x_vector

# save x_vector for model development
with open('data/x_vector.npy', 'wb') as file:
    np.save(file, x_vector)

### Build Terms Matrix

from sklearn.feature_extraction.text import TfidfVectorizer
import scipy.sparse
import json

# create TF-IDF transformer
tfidf = TfidfVectorizer()
tfidf.fit(df.Abstract_Cleaned)

# build terms matrix for entire corpus
terms_sparse_matrix = tfidf.transform(df.Abstract_Cleaned)

# save terms sparse matrix
# to load sparse matrix: sparse_matrix = scipy.sparse.load_npz('data/terms_sparse_matrix.npz')
scipy.sparse.save_npz('data/terms_sparse_matrix.npz', terms_sparse_matrix)

# get term label for each item in the term matrix and save result to a text file
terms_label = tfidf.get_feature_names()

# save term labels
with open("data/terms_label.txt", "w") as fp:
    json.dump(terms_label, fp)
