==> 03_biased_clustering.ipynb <==
# Biased Clustering

This notebook uses time-biased clustering to detect trends in financial journals.

import numpy as np
import pandas as pd
from tqdm import tqdm
from Biased_Clusters import get_clusters_dist, get_clusters_timeline, get_top_keywords, get_silhouette, cal_cluster_bias

# load cleaned data
df = pd.read_csv('data/data_cleaned.csv')
df.head()

## Biased Clustering

# build trend score table

# load training data
x_vector = np.load('data/x_vector.npy')

# steps = [i for i in range(1,1000, 50)]
data = []

# get Silhouette score, std year for various bias amount
for m in tqdm(range(1,1000,10)):
    m = m*.01

    try:
        data.append(get_silhouette(df, x_vector, m))
    except ValueError:
        continue

# create a dataframe to store the results
df_result = pd.DataFrame(data)

# save df_result to csv file
df_result.to_csv('data/table_trend_score.csv', index=False)

# display the first 5 rows of df_result
df_result.head()

import json
import scipy.sparse

# load terms sparse matrix
terms_sparse_matrix = scipy.sparse.load_npz('data/terms_sparse_matrix.npz')

# load terms label
with open("data/terms_label.txt", "r") as fp:
    terms_label = json.load(fp)

# make predictions
summary, predictions = cal_cluster_bias(df, x_vector, terms_sparse_matrix, terms_label, 0.57)

summary

predictions.head()

get_clusters_dist(predictions).head()

# save prediction
summary.to_csv('data/summary_bias_0.57.csv', index=False)
predictions.to_csv('data/predictions_0.57.csv', index=False)
